backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
dj-database-url
django-cors-headers
gunicorn
python-dotenv
django-sendgrid-v5
pytest
pytest-django
pytest-cov
factory-boy
google-api-python-client
requests
drf-spectacular
isbnlib
pre-commit
black
isort
flake8
